+++ core/process.py
@@ -2,7 +2,9 @@ import subprocess
 import platform
 import logging
 
-from data.process_manager import ProcessManager
+import psutil
+
+from data.process_manager import ProcessManager, ProcessPriorityManager
 
 def runProcess2(*cmd: str, cwd: str | None = None) -> (str, str):
     """Replacement for runProcess() and runProcessOutput().
@@ -12,7 +14,8 @@ def runProcess2(*cmd: str, cwd: str | None = None) -> (str, str):
     """
     logging.info(f"[runProcess2] {cmd}")
 
-    process = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, startupinfo=_getStartupInfo(), cwd=cwd)
+    process = psutil.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, startupinfo=_getStartupInfo(), cwd=cwd)
+    _setProcessPriority(process, ProcessPriorityManager.getPriorityFlag())
     ProcessManager.addProcess(process)
     stdout, stderr = process.communicate()
     ProcessManager.removeProcess(process)
@@ -29,6 +32,18 @@ def runProcess2(*cmd: str, cwd: str | None = None) -> (str, str):
         logging.error(f"[runProcess2] Failed to decode process output. {err}")
 
     return (stdout or "", stderr or "")
+
+def _setProcessPriority(process: psutil.Popen, priority: int | None) -> None:
+    """An internal function for setting the priority of a given process."""
+    if priority is None:
+        logging.error(f"[_setProcessPriority] Received None priority, ignoring.")
+        return
+
+    try:
+        process.nice(priority)
+    except (psutil.Error, ValueError) as e:
+        logging.error(f"[_setProcessPriority] Failed to set process priority: {e}")
+        return
 
 def _getStartupInfo():
     """Get startup info for Windows. Prevents console window from showing."""